--- index.ts
@@ -1,10 +1,0 @@
-// export group function and required types
-export { group, Groupable, Collectable, KeyingFunction } from './lib/group'
-
-// export collector results
-export { ArraysCollection } from './lib/collectors/as-arrays'
-export { EntriesCollectorOptions, EntriesCollection } from './lib/collectors/as-entries'
-export { MapCollection } from './lib/collectors/as-map'
-export { ObjectCollection } from './lib/collectors/as-object'
-export { TuplesCollection } from './lib/collectors/as-tuples'
-export { KeysCollection } from './lib/collectors/keys'
