+++ src/collectors/as-arrays.ts
