+++ src/collectors/as-entries.ts
