+++ src/collectors/as-map.ts
