+++ src/collectors/as-object.ts
