+++ src/index.ts
@@ -1,0 +1,10 @@
+// export group function and required types
+export { group, Groupable, Collectable, KeyingFunction } from './group'
+
+// export collector results
+export { ArraysCollection } from './collectors/as-arrays'
+export { EntriesCollectorOptions, EntriesCollection } from './collectors/as-entries'
+export { MapCollection } from './collectors/as-map'
+export { ObjectCollection } from './collectors/as-object'
+export { TuplesCollection } from './collectors/as-tuples'
+export { KeysCollection } from './collectors/keys'
