--- test/.eslintrc.yml
@@ -1,14 +1,0 @@
-root: true
-
-extends:
-  - 'standard-with-typescript'
-
-parserOptions:
-  project: '../tsconfig.eslint.json'
-
-env:
-  mocha: true
-rules:
-  no-unused-expressions: off
-  '@typescript-eslint/no-unused-expressions': off
-  '@typescript-eslint/no-floating-promises': off
