+++ tsconfig.json
@@ -11,9 +11,8 @@
       "./typings"
     ]
   },
-  "exclude": [
-    "test",
-    "dist",
-    "node_modules"
+  "include": [
+    "src",
+    "typings"
   ]
 }
